pyproject.toml
[project]
name = "current-be"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]
